+++ s4e-web/src/app/views/map-view/map.module.ts
@@ -1,3 +1,4 @@
+import { SearchModule } from './../../components/search/search.module';
 import {ModuleWithProviders, NgModule} from '@angular/core';
 import {MapViewComponent} from './map-view.component';
 import {ViewManagerComponent} from './view-manager/view-manager.component';
@@ -15,7 +16,6 @@ import {LegendComponent} from './legend/legend.component';
 import {LegendStore} from './state/legend/legend.store';
 import {LegendQuery} from './state/legend/legend.query';
 import {LegendService} from './state/legend/legend.service';
-import {SearchResultsComponent} from './search-results/search-results.component';
 import {AkitaGuidService} from './state/search-results/guid.service';
 import {OwlDateTimeModule, OwlNativeDateTimeModule} from 'ng-pick-datetime';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
@@ -51,7 +51,6 @@ import { CommonModule } from '@angular/common';
     TimelineComponent,
     MapComponent,
     LegendComponent,
-    SearchResultsComponent,
     ReportModalComponent,
     SentinelSearchComponent,
     ShareConfigurationModalComponent,
@@ -91,6 +90,7 @@ import { CommonModule } from '@angular/common';
             }
         ]),
         S4EFormsModule,
+        SearchModule
         // OwlNativeDateTimeModule
     ],
   providers: [