+++ s4e-web/src/app/views/map-view/state/location-search-results/location-search-results.query.spec.ts
@@ -1,0 +1,14 @@
+import {LocationSearchResultsStore} from './locations-search-results.store';
+import { LocationSearchResultsQuery } from './location-search-results.query';
+
+describe('LocationSearchResultsQuery', () => {
+  let query: LocationSearchResultsQuery;
+
+  beforeEach(() => {
+    query = new LocationSearchResultsQuery(new LocationSearchResultsStore());
+  });
+
+  it('should create an instance', () => {
+    expect(query).toBeTruthy();
+  });
+});