+++ s4e-web/src/app/views/map-view/state/location-search-results/locations-search-results.service.ts
@@ -1,21 +1,21 @@
 import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
-import {SearchResultsStore} from './search-results.store';
-import {SearchResult} from './search-result.model';
+import {LocationSearchResultsStore} from './locations-search-results.store';
+import {LocationSearchResult} from './location-search-result.model';
 import {delay, finalize, map} from 'rxjs/operators';
 import {S4eConfig} from '../../../../utils/initializer/config.service';
-import {AkitaGuidService} from './guid.service';
 import {PageSearchResult} from '../../../../utils/state.types';
 import {MapService} from '../map/map.service';
 import proj4 from 'proj4';
 import {ViewPosition} from '../map/map.model';
-import {SearchResultsQuery} from './search-results.query';
+import { AkitaGuidService } from '../search-results/guid.service';
+import { LocationSearchResultsQuery } from './location-search-results.query';
 
 @Injectable({providedIn: 'root'})
 export class SearchResultsService {
 
-  constructor(private store: SearchResultsStore,
-              private query: SearchResultsQuery,
+  constructor(private store: LocationSearchResultsStore,
+              private query: LocationSearchResultsQuery,
               private guidGenerationService: AkitaGuidService,
               private mapService: MapService,
               private http: HttpClient,
@@ -50,20 +50,16 @@ export class SearchResultsService {
     return ZOOM_LEVELS[type] || null;
   }
 
-  setSelectedPlace(searchResult: SearchResult | null) {
-    this.mapService.setView({
-      centerCoordinates: proj4(
-        this.CONFIG.projection.toProjection,
-        [searchResult.longitude, searchResult.latitude]
-      ),
-      zoomLevel: this.getZoomLevel(searchResult.type)
-    } as ViewPosition);
-    this.store.update({selectedLocation: searchResult, isOpen: false});
-  }
-
-  setFirstAsSelectedPlace() {
-    if(this.query.getAll().length) {
-      this.setSelectedPlace(this.query.getAll()[0]);
+  setSelectedPlace(searchResult: LocationSearchResult | null) {
+    if (!!searchResult) {
+      this.mapService.setView({
+        centerCoordinates: proj4(
+          this.CONFIG.projection.toProjection,
+          [searchResult.longitude, searchResult.latitude]
+        ),
+        zoomLevel: this.getZoomLevel(searchResult.type)
+      } as ViewPosition);
     }
+    this.store.update({searchResult: searchResult, isOpen: false});
   }
 }