+++ s4e-web/src/app/views/map-view/state/location-search-results/locations-search-results.store.spec.ts
@@ -1,0 +1,14 @@
+import {LocationSearchResultsStore} from './locations-search-results.store';
+
+describe('LocationSearchResultsStore', () => {
+  let store: LocationSearchResultsStore;
+
+  beforeEach(() => {
+    store = new LocationSearchResultsStore();
+  });
+
+  it('should create an instance', () => {
+    expect(store).toBeTruthy();
+  });
+
+});