+++ s4e-web/src/app/views/settings/dashboard/dashboard.component.spec.ts
@@ -1,12 +1,22 @@
+import { TileComponent } from './../../../components/tiles-dashboard/tile/tile.component';
 import { DashboardModule } from './dashboard.module';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { async, ComponentFixture, TestBed, tick, fakeAsync } from '@angular/core/testing';
 
 import { DashboardComponent } from './dashboard.component';
 import {RouterTestingModule} from '@angular/router/testing';
 import { DebugElement } from '@angular/core';
-import { Router } from '@angular/router';
+import { Router, convertToParamMap, ParamMap, ActivatedRoute, ActivatedRouteSnapshot } from '@angular/router';
 import { By } from '@angular/platform-browser';
 import { CommonModule } from '@angular/common';
+import { Subject, ReplaySubject } from 'rxjs';
+
+class ActivatedRouteStub {
+  queryParamMap: Subject<ParamMap> = new ReplaySubject(1);
+
+  constructor() {
+    this.queryParamMap.next(convertToParamMap({}));
+  }
+}
 
 describe('DashboardComponent', () => {
   let component: DashboardComponent;
@@ -14,6 +24,7 @@ describe('DashboardComponent', () => {
   let de: DebugElement;
   let router: Router;
   let spy: any;
+  let activatedRoute: ActivatedRouteStub;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
@@ -21,6 +32,9 @@ describe('DashboardComponent', () => {
         CommonModule,
         RouterTestingModule.withRoutes([]),
         DashboardModule
+      ],
+      providers: [
+        {provide: ActivatedRoute, useClass: ActivatedRouteStub}
       ]
     })
     .compileComponents();
@@ -29,6 +43,7 @@ describe('DashboardComponent', () => {
     component = fixture.componentInstance;
     de = fixture.debugElement;
     router = TestBed.get(Router);
+    activatedRoute = <ActivatedRouteStub>TestBed.get(ActivatedRoute);
     spy = spyOn(router, 'navigate').and.stub();
 
     fixture.detectChanges();
@@ -37,4 +52,32 @@ describe('DashboardComponent', () => {
   it('should create', () => {
     expect(component).toBeTruthy();
   });
+
+  it('should enable buttons on known institution', fakeAsync( () => {
+    activatedRoute.queryParamMap.next(convertToParamMap({ institution: 'test '}));
+    tick();
+    fixture.detectChanges();
+
+    const tiles = de.queryAll(By.directive(TileComponent));
+    for (let tile of tiles) {
+      const groupsBtn = tile.query(By.css('.panel__footer .button'));
+      if (groupsBtn) {
+        expect(groupsBtn.nativeElement.disabled).toBeFalsy();
+      }
+    }
+  }));
+
+  it('should disable buttons when institution param is not known', fakeAsync( () => {
+    activatedRoute.queryParamMap.next(convertToParamMap({}));
+    tick();
+    fixture.detectChanges();
+
+    const tiles = de.queryAll(By.directive(TileComponent));
+    for (let tile of tiles) {
+      const groupsBtn = tile.query(By.css('.panel__footer .button'));
+      if (groupsBtn) {
+        expect(groupsBtn.nativeElement.disabled).toBeTruthy();
+      }
+    }
+  }));
 });