+++ s4e-web/src/app/views/settings/state/institution/institution.service.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from '@angular/core/testing';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { InstitutionService } from './institution.service';
 import { InstitutionStore } from './institution.store';
-import {TestingConfigProvider} from '../../../app.configuration.spec';
+import {TestingConfigProvider} from '../../../../app.configuration.spec';
 import {RouterTestingModule} from '@angular/router/testing';
 
 describe('InstitutionService', () => {