+++ s4e-web/src/app/views/settings/state/institution/institution.store.ts
@@ -2,6 +2,7 @@ import {Injectable} from '@angular/core';
 import {ActiveState, EntityState, EntityStore, StoreConfig} from '@datorama/akita';
 import {Institution} from './institution.model';
 
+
 export interface InstitutionState extends EntityState<Institution>, ActiveState<string> {}
 
 @Injectable({providedIn: 'root'})