+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The most recent changes are on top, in each type of changes category.
 
 ### Added
 
+- Add institution selection into super Admin and Admin settings page [#455](https://github.com/cyfronet-fid/sat4envi/issues/455)
+
 ### Changed
 
 ### Fixed